fix: fixed requested changes
